Tests: move Tuple to separate file (.NET 3.5 support)
